pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lewam"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.2.1,<2.11.0",
    "torchvision>=0.21.0,<0.26.0",
    "torchcodec>=0.4.0,<0.11.0",
    "transformers>=5.3",
    "timm>=1.0",
    "einops>=0.8,<0.9.0",
    "accelerate>=1.10.0,<2.0.0",
    "numpy>=2.0.0,<2.3.0",
    "opencv-python>=4.9.0,<4.13.0",
    "pillow>=11.0",
    "matplotlib>=3.10",
    "scikit-learn>=1.8",
    "scipy>=1.17",
    "huggingface-hub>=1.0.0,<2.0.0",
    "safetensors>=0.7",
    "tokenizers>=0.22",
    "av>=15.0.0,<16.0.0",
    "lerobot[feetech] @ git+https://github.com/huggingface/lerobot.git@07502868e58095b437e5dd5a480fecc65a6f29dc",
    "num2words==0.5.14"
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::FutureWarning:timm",
]
